+++ mix/test.cldr.ex
@@ -3,5 +3,6 @@ defmodule Money.Cldr do
 
   use Cldr,
     locales: ["en", "de", "it", "es", "fr"],
+    default_locale: "en",
     providers: [Cldr.Number]
 end